MIM-42 拆分工作台侧栏组件
